Update migrations and adjust schema properties
This commit updates the SQL migrations with a new file and deletes an older one. The database schema experiences changes such as introduction of "address" in "client" table, adjustment of "price" data type in "product" table, setting "client_id" as a unique field in "cart" table, modulating relations in "buy" model, among others. Changes in a Hook method in 'useFetch.tsx' and 'sqldialects.xml' are also included.